progress on ensuring Java's long is represented as int64 (for #83)

+++ get_conflicts_result.go
@@ -1,11 +1,13 @@
 package ravendb
 
+// GetConflictsResult represents result of "get conflict" command
 type GetConflictsResult struct {
 	ID          string      `json:"Id"`
 	Results     []*Conflict `json:"Results"`
-	LargestEtag int         `json:"LargestEtag"`
+	LargestEtag int64       `json:"LargestEtag"`
 }
 
+// Conflict represents conflict
 type Conflict struct {
 	LastModified Time                   `json:"LastModified"`
 	ChangeVector string                 `json:"ChangeVector"`
