+++ hi_lo_result.go
@@ -3,9 +3,9 @@ package ravendb
 // HiLoResult is a result of HiLoResult command
 type HiLoResult struct {
 	Prefix      string `json:"Prefix"`
-	Low         int    `json:"Low"`
-	High        int    `json:"High"`
-	LastSize    int    `json:"LastSize"`
+	Low         int64  `json:"Low"`
+	High        int64  `json:"High"`
+	LastSize    int64  `json:"LastSize"`
 	ServerTag   string `json:"ServerTag"`
 	LastRangeAt *Time  `json:"LastRangeAt"`
 }
